+++ frontend/src/App.tsx
@@ -1,8 +1,7 @@
-import { useEffect, useState, VFC } from 'react'
+import { useState, VFC } from 'react'
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
 
 import { ChakraProvider } from '@chakra-ui/react'
-import firebase from 'firebase/compat/app'
 import { DateTime } from 'luxon'
 
 import { EmailConfirmation } from 'components/pages/EmailConfirmation'
@@ -13,21 +12,14 @@ import { Setting } from 'components/pages/Setting'
 import { ShowPaymentEntry } from 'components/pages/ShowPaymentEntry'
 import { SignIn } from 'components/pages/SignIn'
 import { PrivateRoute } from 'components/router/PrivateRoute'
-import { AuthContext } from 'context/AuthContext'
 import { PaymentContext } from 'context/PaymentContext'
-import { getCurrentUser } from 'lib/api/auth'
-import { auth } from 'lib/firebase'
+import { AuthProvider } from 'provider/AuthProvider'
 import { theme } from 'theme/index'
-import { GetCurrentUserParams } from 'types/getCurrentUserParams'
 
 import type { PaymentListGroupByPaidAt } from 'types/paymentListGroupByPaidAt'
 import type { TeamStatus } from 'types/teamStatus'
-import type { User } from 'types/user'
 
 const App: VFC = () => {
-  const [isLoaded, setIsLoaded] = useState<boolean>(false)
-  const [currentUser, setCurrentUser] = useState<User | null>(null)
-  const [currentFirebaseUser, setCurrentFirebaseUser] = useState<firebase.User | null>(null)
   const [paymentList, setPaymentList] = useState<PaymentListGroupByPaidAt[]>([])
   const [teamStatus, setTeamStatus] = useState<TeamStatus>({
     refundAmount: 0,
@@ -42,57 +34,14 @@ const App: VFC = () => {
   const [inputDetail, setInputDetail] = useState<string>('')
   const [inputPaidAt, setInputPaidAt] = useState<string>(DateTime.local().toFormat('yyyy-MM-dd'))
 
-  const handleGetCurrentUser = async () => {
-    const token = await auth.currentUser?.getIdToken(true)
-    const params: GetCurrentUserParams = {
-      uid: auth.currentUser?.uid,
-    }
-    try {
-      const res = await getCurrentUser(params, token)
-      if (res?.status === 200) {
-        setCurrentUser(res.data.user)
-      }
-    } catch (e) {
-      console.log(e)
-    } finally {
-      setIsLoaded(true)
-    }
-  }
-
-  useEffect(() => {
-    const unsubscribed = auth.onAuthStateChanged((user) => {
-      setCurrentFirebaseUser(user)
-    })
-    return () => {
-      unsubscribed()
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  useEffect(() => {
-    handleGetCurrentUser().catch((err) => {
-      console.log(err)
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentFirebaseUser])
-
   return (
     <BrowserRouter>
       <ChakraProvider theme={theme}>
-        <AuthContext.Provider
-          // eslint-disable-next-line react/jsx-no-constructed-context-values
-          value={{
-            isLoaded,
-            setIsLoaded,
-            currentUser,
-            setCurrentUser,
-            currentFirebaseUser,
-          }}
-        >
-          <Switch>
-            <Route path="/signin" component={SignIn} />
-            <Route path="/invitation" component={Invitation} />
-            <Route path="/confirmation" component={EmailConfirmation} />
+        <Switch>
+          <Route path="/signin" component={SignIn} />
+          <Route path="/invitation" component={Invitation} />
+          <Route path="/confirmation" component={EmailConfirmation} />
+          <AuthProvider>
             <PaymentContext.Provider
               // eslint-disable-next-line react/jsx-no-constructed-context-values
               value={{
@@ -126,8 +75,8 @@ const App: VFC = () => {
                 />
               </PrivateRoute>
             </PaymentContext.Provider>
-          </Switch>
-        </AuthContext.Provider>
+          </AuthProvider>
+        </Switch>
       </ChakraProvider>
     </BrowserRouter>
   )
